requirements.txt
pandas>=1.3.0
openpyxl>=3.0.0
python-docx>=0.8.11
PyPDF2>=3.0.0
docx2pdf>=0.1.8